+++ packages/jsenv-plugin-toolbar/src/client/core/execution_actions.js
@@ -1,0 +1,9 @@
+import { executionTooltipRequestedSignal } from "./execution_signals.js"
+
+export const requestExecutionTooltip = () => {
+  executionTooltipRequestedSignal.value = true
+}
+
+export const closeExecutionTooltip = () => {
+  executionTooltipRequestedSignal.value = false
+}
