+++ packages/jsenv-plugin-toolbar/src/client/core/execution_signals.js
@@ -1,0 +1,3 @@
+import { signal } from "@preact/signals"
+
+export const executionTooltipRequestedSignal = signal(false)
