+++ packages/jsenv-plugin-toolbar/src/client/core/server_actions.js
@@ -1,0 +1,9 @@
+import { serverTooltipRequestedSignal } from "./server_signals.js"
+
+export const requestServerTooltip = () => {
+  serverTooltipRequestedSignal.value = true
+}
+
+export const closeServerTooltip = () => {
+  serverTooltipRequestedSignal.value = false
+}
