+++ packages/jsenv-plugin-toolbar/src/client/toolbar.html
@@ -163,7 +163,7 @@
               <svg>
                 <circle
                   class="server_indicator_circle"
-                  data-variant="off"
+                  data-variant="closed"
                   cx="20"
                   cy="20"
                   r="5"
